refactor-emit-raw
Move entire raw listener function to separate file
Change the switch statements to a object
Remove the raw listener/event all together since it's not needed and
isn't exactly 'raw'

Easier for testing, and maintaining
